DL/pretrain/requirements.txt
fastapi
uvicorn
pydantic
Pillow
torch
torchvision
transformers
